+++ C/42/libft/ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: Nissya <kirito042004@gmail.com>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/13 18:46:33 by Nissya            #+#    #+#             */
-/*   Updated: 2023/11/13 19:12:17 by Nissya           ###   ########.fr       */
+/*   Updated: 2023/11/30 10:45:22 by Nissya           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dest, const void *src, int n)
+#include "libft.h"
+
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	const unsigned char	*src_ptr;
 	int					i;
