+++ C/42/libft/ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: Nissya <kirito042004@gmail.com>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/13 18:16:55 by Nissya            #+#    #+#             */
-/*   Updated: 2023/11/14 11:07:07 by Nissya           ###   ########.fr       */
+/*   Updated: 2023/11/30 10:48:14 by Nissya           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *dest, int ch, int n)
+#include "libft.h"
+
+void	*ft_memset(void *dest, int ch, size_t n)
 {
 	int				i;
 	unsigned char	*ptr;
